requirements.txt
absl-py==0.11.0
antlr4-python3-runtime==4.8
appnope==0.1.2
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.2.1
cached-property==1.5.2
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
colorama==0.4.4
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
fsspec==0.8.4
future==0.18.2
google-auth-oauthlib==0.4.2
google-auth==1.24.0
grpcio==1.34.0
h5py==3.1.0
hydra-core==1.0.4
idna==2.10
imageio==2.9.0
importlib-metadata==3.2.0
importlib-resources==3.3.0
ipydatawidgets==4.1.0
ipykernel==5.4.2
ipython-genutils==0.2.0
ipython==7.19.0
ipywidgets==7.5.1
jedi==0.17.2
jinja2==2.11.2
joblib==0.17.0
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyter==1.0.0
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
jupyterlab==2.2.9
k3d==2.9.1
kiwisolver==1.3.1
markdown==3.3.3
markupsafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
msgpack==1.0.1
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
notebook==6.1.5
numpy==1.19.4
oauthlib==3.1.0
omegaconf==2.0.5
packaging==20.8
pandas==1.1.5
pandocfilters==1.4.3
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
pillow==8.0.1
prometheus-client==0.9.0
prompt-toolkit==3.0.8
protobuf==3.14.0
ptyprocess==0.6.0
py==1.10.0
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.20
pygments==2.7.3
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytorch-lightning==1.1.0
pytz==2020.4
pywinpty==0.5.7
pyyaml==5.3.1
pyzmq==20.0.0
qtconsole==5.0.1
qtpy==1.9.0
requests-oauthlib==1.3.0
requests==2.25.0
rsa==4.6
scikit-learn==0.23.2
scipy==1.5.4
send2trash==1.5.0
six==1.15.0
tabulate==0.8.7
tensorboard-plugin-wit==1.7.0
tensorboard==2.4.0
terminado==0.9.1
test-tube==0.7.5
testpath==0.4.4
threadpoolctl==2.1.0
timm==0.3.2
torch==1.6.0
torchvision==0.7.0
tornado==6.1
tqdm==4.54.1
traitlets==5.0.5
traittypes==0.2.1
typing-extensions==3.7.4.3
urllib3==1.26.2
wcwidth==0.2.5
webencodings==0.5.1
werkzeug==1.0.1
widgetsnbextension==3.5.1
zipp==3.4.0
